feat(update): implement background update and improve UI
- Add BUTTON_INSTALL translation key
- Update TITLE_DOWNLOADING_NEW_VERSION to include newVersion parameter
- Implement background download and install process
- Improve update dialog UI with progress bar, release notes, and install button
- Add error handling and toast notification for update failures
